перенес хуки из инпута
