--- src/components/hooks/formValidation.jsx
@@ -1,2 +1,0 @@
-import { useState, useRef } from "react";
-import { regularEmail } from "../../services/consts";
