+++ src/services/consts.jsx
@@ -1,2 +1,5 @@
-export const regularEmail =
-  /^([A-Za-z0-9_\.-]+)@([a-z0-9_\.-]+)\.([a-z]{2,3})$/;
+const regularEmail = /^([A-Za-z0-9_\.-]+)@([a-z0-9_\.-]+)\.([a-z]{2,3})$/;
+
+const regularName = /[/!$()*+.<>?^{|}_0-9]/;
+
+export { regularName, regularEmail };
